test(canon): split the documented-difference typecheck divergence tests
